+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Bump pywebpush from 2.4.0 to 2.5.0
 - Adjusting the search page for a more consistent design
 - Adjusting invitations pages for a more consistent design
+- Adjusting profile settings page for a more consistent design
 
 ## Release 0.8.3
 - Dependency update: cffi 2.1.1